+++ stac_fastapi/extensions/stac_fastapi/extensions/core/pagination/__init__.py
@@ -1,6 +1,7 @@
 """Pagination classes as extensions."""
 
+from .offset_pagination import OffsetPaginationExtension
 from .pagination import PaginationExtension
 from .token_pagination import TokenPaginationExtension
 
-__all__ = ["PaginationExtension", "TokenPaginationExtension"]
+__all__ = ["OffsetPaginationExtension", "PaginationExtension", "TokenPaginationExtension"]